Adição ParticipanteRaid e algumas implementações de segurança

+++ src/main/java/com/pi4/dungeonFinderBackend/services/RaidService.java
@@ -3,7 +3,9 @@ package com.pi4.dungeonFinderBackend.services;
 import com.pi4.dungeonFinderBackend.datasource.repositories.JogoRepository;
 import com.pi4.dungeonFinderBackend.datasource.repositories.RaidRepository;
 import com.pi4.dungeonFinderBackend.datasource.repositories.UsuarioRepository;
+import com.pi4.dungeonFinderBackend.domain.entities.Jogo;
 import com.pi4.dungeonFinderBackend.domain.entities.Raid;
+import com.pi4.dungeonFinderBackend.domain.entities.Usuario;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -21,8 +23,12 @@ public class RaidService {
     private final JogoRepository jogoRepository;
     private final UsuarioRepository usuarioRepository;
 
-    public List<Raid> listarTodos(UUID id) {
+    public List<Raid> listarTodos() {
         return raidRepository.findAll();
+    }
+
+    public List<Raid> listarPorJogo(UUID jogoId) {
+        return raidRepository.findByJogoId(jogoId);
     }
 
     public Raid buscarPorId(UUID id) {
@@ -30,27 +36,28 @@ public class RaidService {
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Raid não encontrada"));
     }
 
-    public Raid criar(Raid raid, UUID id) {
-
+    public Raid criar(Raid raid, UUID jogoId, UUID usuarioId) {
         if (raidRepository.existsByNome(raid.getNome()))
             throw new ResponseStatusException(HttpStatus.CONFLICT, "Nome já cadastrado");
 
+        Jogo jogo = jogoRepository.findById(jogoId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Jogo não encontrado"));
+
+        Usuario usuario = usuarioRepository.findById(usuarioId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário não encontrado"));
+
+        raid.setJogo(jogo);
+        raid.setCriadoPor(usuario);
         raid.setCriadoEm(LocalDateTime.now());
         return raidRepository.save(raid);
     }
 
-    public Raid atualizar(UUID id, Raid dadosNovos) {
-
+    public void deletar(UUID id, UUID usuarioId) {
         Raid raid = buscarPorId(id);
-        raid.setNome(dadosNovos.getNome());
-        raid.setDescricao(dadosNovos.getDescricao());
 
-        return raidRepository.save(raid);
-    }
+        if (!raid.getCriadoPor().getId().equals(usuarioId))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Apenas o criador pode deletar esta raid");
 
-    public void deletar(UUID id) {
-        buscarPorId(id);
         raidRepository.deleteById(id);
     }
-
 }
